Added website_sale_coupon Odoo module removal in migration script
Since this Odoo module carries the same name as one of Commown's modules,
its presence in the modules list leads to the Odoo module and its dependencies
being installed, as well as its migrations scripts being ran,
leading to migration crashes.

As such, we remove the Odoo version of the website_sale_coupon module,
which we don't use anyway.