[OpenUpgrade 16.0] Added branch fork removing a module rename
The website_sale_coupon module is marked as renamed in the OpenUpgrade apriori.py file,
since the Odoo module is being renamed to a new name.
However, since this conflicts with our custom module carrying the same name,
this renaming can lead to installation of additionnal unwanted modules,
as well as not installing the wanted module.

As such, we forked the branch to comment the dict entry, which will avoid this rename.