[step 4 (14->15)] Added request to assign newly created payment method lines to unassigned account payments
This query is taken from the account 15.0 end-migration.py script, which assigns an account.payment.method.line id
to account.payment records.

We execute it again since our account.payment.method.line records are created in the payment end-migration script,
which is only executed after the account script, and as such, the newly created lines are not assigned.

(See ticket #44773 for more info)