[step 6 (16.0 final.)] Fixed print label group assignment on users roles
Using a single recordset comprising of all of the user roles, then accessing all of their implied groups at once with an union assignment,
lead to the assignment of all the groups of each indivudual role to all of the roles.

As such, we instead assign the group to each role individually in a loop, to avoid this buggy union.

Note, I considered replacing the union assignment by a Command link ([4, label_group.id, 0]) update instruction,
but I feel this form is easier to parse and read.