[step 5 (15.0 -> 16.0)] Set overloaded translatable fields to True before migration
In our modules, these fields are inherited from other models, which we modify to be translatable.
However, between Odoo v12 and v16, these fields still exist, and are set to translate=False
before the modules are installed again.

This is problematic, as in the pre-migration Openupgrade phase of the base module, the translatable fields
are migrated from the ir_translation tables to being stored within the table columns directly, and this procedure
happens for translatable fields (translate == True).

As such, to insure these fields' translations are properly migrated, we set these fields to translated,
as they would have been if we simply readded these modules during the migration, but allows us to benefit
from the translations migration.